+++ .env.sample
@@ -1,4 +1,6 @@
 DATABASE_URL=file:./transcendence_dev.sqlite3
-REDIS_URL=redis://localhost:6379
+REDIS_URL=redis://redis:6379
 GRAFANA_USER=admin
 GRAFANA_PASS=admin
+OTEL_SERVICE_NAME=fastify-app
+OTEL_METRICS_PORT=9464
